requirements.txt
asgiref==3.4.1
attrs==21.2.0
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
colorama==0.4.4
cryptography==3.4.7
distro==1.6.0
Django==3.2.6
docker==5.0.0
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
idna==3.2
jsonschema==3.2.0
paramiko==2.7.2
pipreqs==0.4.10
pycparser==2.20
PyNaCl==1.4.0
pyrsistent==0.18.0
python-dotenv==0.19.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
texttable==1.6.4
urllib3==1.26.6
websocket-client==0.59.0
yarg==0.1.9
psycopg2-binary==2.9
Pillow==8.3.0
djangorestframework==3.12.2
django-rest-auth==0.9.5
django-allauth==0.45.0
invoke==1.6.0